Removing the request response object to utils folder
